Added different animation for start button
There are 2 button animations that I think would look good here. I've put the pressing animation for the start button and the sliding animation for the other two.
If you want to check the sliding animation for the start button check the previous commit.
Also tell me which animation I should apply for all the buttons. Im in the favour of the pressing kind